Continious Deployment3

+++ Jenkinsfile
@@ -28,9 +28,12 @@ pipeline {
         stage('Deploy to Docker') {
             steps {
                 script {
-                    bat 'docker pull roshanteli/devops-integration:latest'
-                    bat 'docker stop devops-integration || true'
-                    bat 'docker rm devops-integration || true'
+          			def containerExists = bat(script: 'docker ps -a --filter "name=devops-integration" --format "{{.Names}}"', returnStdout: true).trim()
+
+                    if (containerExists) {
+                        bat 'docker stop devops-integration || true'
+                        bat 'docker rm devops-integration || true'
+                    }
                     bat 'docker run -d --name devops-integration -p 8082:8080 roshanteli/devops-integration:latest'
                 }
             }
